Split up in separate classes using autoploader
