+++ src/HTMLForm/CFormElementColor.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Mos\HTMLForm;
+
+/**
+ * Form element for color
+ */
+class CFormElementColor extends CFormElement
+{
+
+    /**
+     * Constructor
+     *
+     * @param string $name       of the element.
+     * @param array  $attributes to set to the element. Default is an empty array.
+     *
+     * @return void
+     */
+    public function __construct($name, $attributes = []) 
+    {
+        parent::__construct($name, $attributes);
+        $this['type'] = 'color';
+        $this->UseNameAsDefaultLabel();
+    }
+}
