+++ src/HTMLForm/CFormElementFileMultiple.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Mos\HTMLForm;
+
+/**
+ * Form element 
+ */
+class CFormElementFileMultiple extends CFormElement
+{
+    /**
+     * Constructor
+     *
+     * @param string $name       of the element.
+     * @param array  $attributes to set to the element. Default is an empty array.
+     *
+     * @return void
+     */
+    public function __construct($name, $attributes = []) 
+    {
+        parent::__construct($name, $attributes);
+        $this['type'] = 'file-multiple';
+        $this->UseNameAsDefaultLabel();
+    }
+}
