+++ webroot/test/test-elements-html401.php
@@ -1,6 +1,6 @@
 <?php
 // Include CForm
-include('../CForm.php');
+include('../../autoloader.php');
 
 
 // -----------------------------------------------------------------------
@@ -9,7 +9,7 @@ include('../CForm.php');
 //
 session_name('cform_example');
 session_start();
-$form = new CForm(array('legend' => 'Legend'), array(
+$form = new \Mos\HTMLForm\CForm(array('legend' => 'Legend'), array(
     'text' => array(
       'type'        => 'text',
       'description' => 'Here you can place a description.',
@@ -121,4 +121,4 @@ else if($status === false){
 <h1>CForm Example: Form elements as of HTML 4.01</h1>
 <?=$form->GetHTML()?>
 
-<?php $footer = "../../template/footer_mos.php"; if(is_file($footer)) include($footer) ?>
+<?php $footer = "footer_mos.php"; if(is_file($footer)) include($footer) ?>
